# shared-contracts/target-operational-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "EnterpriseSitePage": {
        "additionalProperties": false,
        "properties": {
          "evaluatedAt": {
            "format": "date-time",
            "title": "Evaluatedat",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EnterpriseSiteResource"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor",
          "evaluatedAt"
        ],
        "title": "EnterpriseSitePage",
        "type": "object"
      },
      "EnterpriseSiteResource": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "barangay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barangay"
          },
          "buildingCapacity": {
            "title": "Buildingcapacity",
            "type": "integer"
          },
          "classification": {
            "enum": [
              "official",
              "simulation"
            ],
            "title": "Classification",
            "type": "string"
          },
          "coordinatesUpdatedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinatesupdatedat"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/SiteDeviceResource"
            },
            "title": "Devices",
            "type": "array"
          },
          "enterpriseCategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprisecategory"
          },
          "enterpriseCode": {
            "title": "Enterprisecode",
            "type": "string"
          },
          "enterpriseId": {
            "format": "uuid",
            "title": "Enterpriseid",
            "type": "string"
          },
          "enterpriseLifecycleState": {
            "enum": [
              "active",
              "inactive"
            ],
            "title": "Enterpriselifecyclestate",
            "type": "string"
          },
          "enterpriseName": {
            "title": "Enterprisename",
            "type": "string"
          },
          "geocodedAddress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geocodedaddress"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "liveState": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SiteLiveStateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "locationVersion": {
            "minimum": 1.0,
            "title": "Locationversion",
            "type": "integer"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "siteCode": {
            "title": "Sitecode",
            "type": "string"
          },
          "siteId": {
            "format": "uuid",
            "title": "Siteid",
            "type": "string"
          },
          "siteName": {
            "title": "Sitename",
            "type": "string"
          },
          "timezone": {
            "const": "Asia/Manila",
            "title": "Timezone",
            "type": "string"
          },
          "topologyStatus": {
            "enum": [
              "ready",
              "unlinked",
              "ambiguous"
            ],
            "title": "Topologystatus",
            "type": "string"
          }
        },
        "required": [
          "siteId",
          "enterpriseId",
          "enterpriseCode",
          "enterpriseName",
          "enterpriseCategory",
          "enterpriseLifecycleState",
          "classification",
          "siteCode",
          "siteName",
          "barangay",
          "address",
          "geocodedAddress",
          "latitude",
          "longitude",
          "buildingCapacity",
          "timezone",
          "locationVersion",
          "coordinatesUpdatedAt",
          "topologyStatus",
          "devices",
          "liveState"
        ],
        "title": "EnterpriseSiteResource",
        "type": "object"
      },
      "EpochAcknowledgementResource": {
        "additionalProperties": false,
        "properties": {
          "counterEpoch": {
            "format": "uuid",
            "title": "Counterepoch",
            "type": "string"
          },
          "deviceId": {
            "format": "uuid",
            "title": "Deviceid",
            "type": "string"
          },
          "epochGeneration": {
            "minimum": 1.0,
            "title": "Epochgeneration",
            "type": "integer"
          },
          "telemetryEpochId": {
            "format": "uuid",
            "title": "Telemetryepochid",
            "type": "string"
          }
        },
        "required": [
          "telemetryEpochId",
          "deviceId",
          "counterEpoch",
          "epochGeneration"
        ],
        "title": "EpochAcknowledgementResource",
        "type": "object"
      },
      "EpochStartAcknowledgement": {
        "additionalProperties": false,
        "properties": {
          "acknowledgedAt": {
            "format": "date-time",
            "title": "Acknowledgedat",
            "type": "string"
          },
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "disposition": {
            "enum": [
              "created",
              "replayed"
            ],
            "title": "Disposition",
            "type": "string"
          },
          "payloadHash": {
            "title": "Payloadhash",
            "type": "string"
          },
          "resource": {
            "$ref": "#/components/schemas/EpochAcknowledgementResource"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "disposition",
          "payloadHash",
          "acknowledgedAt",
          "resource"
        ],
        "title": "EpochStartAcknowledgement",
        "type": "object"
      },
      "EpochStartCommand": {
        "additionalProperties": false,
        "properties": {
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "expectedVersion": {
            "minimum": 0.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "idempotencyKey": {
            "maxLength": 240,
            "minLength": 8,
            "title": "Idempotencykey",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/EpochStartPayload"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "idempotencyKey",
          "occurredAt",
          "expectedVersion",
          "payload"
        ],
        "title": "EpochStartCommand",
        "type": "object"
      },
      "EpochStartPayload": {
        "additionalProperties": false,
        "properties": {
          "counterEpoch": {
            "format": "uuid",
            "title": "Counterepoch",
            "type": "string"
          },
          "deviceId": {
            "format": "uuid",
            "title": "Deviceid",
            "type": "string"
          },
          "expectedPreviousEpoch": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedpreviousepoch"
          }
        },
        "required": [
          "deviceId",
          "counterEpoch",
          "expectedPreviousEpoch"
        ],
        "title": "EpochStartPayload",
        "type": "object"
      },
...
      "FleetSimulationCommand": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "elapsedSeconds": {
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Elapsedseconds",
            "type": "integer"
          },
          "runId": {
            "maxLength": 80,
            "minLength": 3,
            "title": "Runid",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "title": "Startedat",
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/FleetSimulationTarget"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "contractVersion",
          "runId",
          "startedAt",
          "elapsedSeconds",
          "targets"
        ],
        "title": "FleetSimulationCommand",
        "type": "object"
      },
      "FleetSimulationEnterprise": {
        "additionalProperties": false,
        "properties": {
          "barangay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barangay"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "enterpriseId": {
            "title": "Enterpriseid",
            "type": "string"
          },
          "enterpriseName": {
            "title": "Enterprisename",
            "type": "string"
          }
        },
        "required": [
          "enterpriseId",
          "enterpriseName",
          "category",
          "barangay"
        ],
        "title": "FleetSimulationEnterprise",
        "type": "object"
      },
      "FleetSimulationResult": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/TelemetryAcknowledgement"
            },
            "title": "Observations",
            "type": "array"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          }
        },
        "required": [
          "contractVersion",
          "runId",
          "observations"
        ],
        "title": "FleetSimulationResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TelemetryAcknowledgement": {
        "additionalProperties": false,
        "properties": {
          "acknowledgedAt": {
            "format": "date-time",
            "title": "Acknowledgedat",
            "type": "string"
          },
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "disposition": {
            "enum": [
              "created",
              "replayed"
            ],
            "title": "Disposition",
            "type": "string"
          },
          "payloadHash": {
            "title": "Payloadhash",
            "type": "string"
          },
          "resource": {
            "$ref": "#/components/schemas/TelemetryAcknowledgementResource"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "disposition",
          "payloadHash",
          "acknowledgedAt",
          "resource"
        ],
        "title": "TelemetryAcknowledgement",
        "type": "object"
      },
      "TelemetryAcknowledgementResource": {
        "additionalProperties": false,
        "properties": {
          "becameCurrent": {
            "title": "Becamecurrent",
            "type": "boolean"
          },
          "counterEpoch": {
            "format": "uuid",
            "title": "Counterepoch",
            "type": "string"
          },
          "epochGeneration": {
            "minimum": 1.0,
            "title": "Epochgeneration",
            "type": "integer"
          },
          "liveStateVersion": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Livestateversion"
          },
          "observationId": {
            "format": "uuid",
            "title": "Observationid",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "siteId": {
            "format": "uuid",
            "title": "Siteid",
            "type": "string"
          }
        },
        "required": [
          "observationId",
          "siteId",
          "counterEpoch",
          "epochGeneration",
          "sequence",
          "becameCurrent"
        ],
        "title": "TelemetryAcknowledgementResource",
        "type": "object"
      },
      "TelemetryCommand": {
        "additionalProperties": false,
        "properties": {
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "idempotencyKey": {
            "maxLength": 240,
            "minLength": 8,
            "title": "Idempotencykey",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/TelemetryCommandPayload"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "idempotencyKey",
          "occurredAt",
          "expectedVersion",
          "payload"
        ],
        "title": "TelemetryCommand",
        "type": "object"
      },
      "TelemetryCommandPayload": {
        "additionalProperties": false,
        "properties": {
          "counterEpoch": {
            "format": "uuid",
            "title": "Counterepoch",
            "type": "string"
          },
          "deviceHealth": {
            "$ref": "#/components/schemas/DeviceHealth"
          },
          "deviceId": {
            "format": "uuid",
            "title": "Deviceid",
            "type": "string"
          },
          "epochGeneration": {
            "minimum": 1.0,
            "title": "Epochgeneration",
            "type": "integer"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/TelemetryMetric"
            },
            "maxItems": 100,
            "title": "Metrics",
            "type": "array"
          },
          "observedAt": {
            "format": "date-time",
            "title": "Observedat",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "syncHealth": {
            "$ref": "#/components/schemas/SyncHealth"
          }
        },
        "required": [
          "deviceId",
          "counterEpoch",
          "epochGeneration",
          "sequence",
          "observedAt",
          "deviceHealth",
          "syncHealth"
        ],
        "title": "TelemetryCommandPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TANAW API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operational/desktop/sites/v2": {
      "get": {
        "operationId": "get_enterprise_sites_v2_operational_desktop_sites_v2_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "afterSiteId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aftersiteid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseSitePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Enterprise Sites V2",
        "tags": [
          "telemetry-v2"
        ]
      }
    },
    "/operational/desktop/telemetry-epochs/v2": {
      "post": {
        "operationId": "start_telemetry_epoch_v2_operational_desktop_telemetry_epochs_v2_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpochStartCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpochStartAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Telemetry Epoch V2",
        "tags": [
          "telemetry-v2"
        ]
      }
    },
    "/operational/desktop/telemetry/v2": {
      "post": {
        "operationId": "ingest_telemetry_v2_operational_desktop_telemetry_v2_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Telemetry V2",
        "tags": [
          "telemetry-v2"
        ]
      }
    },
...
    "/operational/simulation/sites/v2": {
      "get": {
        "operationId": "list_simulation_sites_v2_operational_simulation_sites_v2_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FleetSimulationEnterprise"
                  },
                  "title": "Response List Simulation Sites V2 Operational Simulation Sites V2 Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Simulation Sites V2",
        "tags": [
          "telemetry-v2"
        ]
      }
    },
    "/operational/simulation/telemetry/v2": {
      "post": {
        "operationId": "ingest_simulation_telemetry_v2_operational_simulation_telemetry_v2_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FleetSimulationCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FleetSimulationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Simulation Telemetry V2",
        "tags": [
          "telemetry-v2"
        ]
      }
    },
    "/operational/sites/v2": {
      "get": {
        "operationId": "get_official_sites_v2_operational_sites_v2_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "afterSiteId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aftersiteid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseSitePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Official Sites V2",
        "tags": [
          "telemetry-v2"
        ]
      }
    },
...
  }
}